send_signal will reload or restart container
Depending on the value of the passed option signal_method,
send_signal will either send a SIGHUP signal to the container, or,
it will pass a restart message to the container.

+++ source/hostel_huptainer/containers.py
@@ -14,8 +14,12 @@ def csv_contains_value(csv_list, value):
 
 
 def send_signal(signal_method, container):
-    """Use the specified method to reload or restart the container."""
-    container.kill(signal='SIGHUP')
+    """Reload or restart the container."""
+    if signal_method == 'reload':
+        container.kill(signal='SIGHUP')
+
+    elif signal_method == 'restart':
+        container.restart()
 
 
 class MatchingContainers(object):
